Fix fornth-child issues caused by empty TRs
When the entirety of a row is shifted up the resulting structure has an empty <tr></tr> pair after the completed row. This causes nth-child based CSS to be, well, wrong if you don't pad to account for it.

This fix wraps <tr></tr> in <tbody></tbody> and drops the empty <tr></tr> sets. This allows the CSS to be set on the <tbody> instead of the <tr> for the desired effects.